add back includeColumnStats to ScanContext and moved SplitHelpers to the same package as ScanContext
